requirements.txt
numpy==1.18.2
scipy==1.4.1
matplotlib==3.2.1
tqdm==4.44.1
scikit-learn==0.22.2.post1
scikit-image==0.16.2
pillow==7.0.0
SimpleITK==1.2.4
tensorflow-gpu==2.1.0
keras==2.3.1
h5py==2.10.0
opencv-python==4.2.0.32